requirements.txt
Django==3.2
djangorestframework==3.12.4
django-redis==4.12.1
celery==5.1.2
gunicorn==20.1.0